+++ src/components/InputTodo/InputTodo.js
@@ -1,11 +1,11 @@
-import { FaPlusCircle, FaSpinner } from "react-icons/fa";
-import { useCallback, useEffect, useState } from "react";
+import { FaPlusCircle, FaSpinner } from 'react-icons/fa';
+import { useCallback, useEffect, useState } from 'react';
 
-import { createTodo } from "../api/todo";
-import useFocus from "../hooks/useFocus";
+import { createTodo } from '../../api/todo';
+import useFocus from '../../hooks/useFocus';
 
 const InputTodo = ({ setTodos }) => {
-  const [inputText, setInputText] = useState("");
+  const [inputText, setInputText] = useState('');
   const [isLoading, setIsLoading] = useState(false);
   const { ref, setFocus } = useFocus();
 
@@ -14,31 +14,31 @@ const InputTodo = ({ setTodos }) => {
   }, [setFocus]);
 
   const handleSubmit = useCallback(
-    async (e) => {
+    async e => {
       try {
         e.preventDefault();
         setIsLoading(true);
 
         const trimmed = inputText.trim();
         if (!trimmed) {
-          return alert("Please write something");
+          return alert('Please write something');
         }
 
         const newItem = { title: trimmed };
         const { data } = await createTodo(newItem);
 
         if (data) {
-          return setTodos((prev) => [...prev, data]);
+          return setTodos(prev => [...prev, data]);
         }
       } catch (error) {
         console.error(error);
-        alert("Something went wrong.");
+        alert('Something went wrong.');
       } finally {
-        setInputText("");
+        setInputText('');
         setIsLoading(false);
       }
     },
-    [inputText, setTodos],
+    [inputText, setTodos]
   );
 
   return (
@@ -48,7 +48,7 @@ const InputTodo = ({ setTodos }) => {
         placeholder="Add new todo..."
         ref={ref}
         value={inputText}
-        onChange={(e) => setInputText(e.target.value)}
+        onChange={e => setInputText(e.target.value)}
         disabled={isLoading}
       />
       {!isLoading ? (
