+++ src/pages/Main.js
@@ -1,9 +1,9 @@
-import { useEffect, useState } from "react";
+import { useEffect, useState } from 'react';
 
-import Header from "../components/Header";
-import InputTodo from "../components/InputTodo";
-import TodoList from "../components/TodoList";
-import { getTodoList } from "../api/todo";
+import Header from '../components/Header/Header';
+import InputTodo from '../components/InputTodo/InputTodo';
+import TodoList from '../components/TodoList/TodoList';
+import { getTodoList } from '../api/todo';
 
 const Main = () => {
   const [todoListData, setTodoListData] = useState([]);
